PGB-1598 Run messagebroker thread when app is started in uwsgi mode as well
