+++ src/gobstuf/wsgi.py
@@ -1,4 +1,4 @@
-from gobstuf.api import get_app
+from gobstuf.app import get_app
 
 # Run the app with uWSGI
 app = get_app()
