+++ src/tests/test_wsgi.py
@@ -3,7 +3,7 @@ from unittest import mock
 
 class TestWsgi(unittest.TestCase):
 
-    @mock.patch('gobstuf.api.get_app')
+    @mock.patch('gobstuf.app.get_app')
     def test_wsgi(self, mock_get_app):
         import gobstuf.wsgi
         mock_get_app.assert_called()
